fix transitions using nvidia accel

+++ ErsatzTV.Application/Streaming/PtsTime.cs
@@ -2,15 +2,15 @@
 
 namespace ErsatzTV.Application.Streaming;
 
-public record PtsTime(double Value)
+public record PtsTime(long Value)
 {
     public static readonly PtsTime Zero = new(0);
 
     public static PtsTime From(string ffprobeLine)
     {
         string[] split = ffprobeLine.Split("|");
-        var ptsTime = double.Parse(split[0], CultureInfo.InvariantCulture);
-        if (double.TryParse(split[1], CultureInfo.InvariantCulture, out double duration))
+        var ptsTime = long.Parse(split[0], CultureInfo.InvariantCulture);
+        if (long.TryParse(split[1], CultureInfo.InvariantCulture, out long duration))
         {
             ptsTime += duration;
         }
